# Remove dependency on Scala parallel collections in toolchain