--- a/tools-benchmarks/src/main/scala/scala/scalanative/benchmarks/testinterface/CodeGenBench.scala
+++ b/tools-benchmarks/src/main/scala/scala/scalanative/benchmarks/testinterface/CodeGenBench.scala
@@ -10,13 +10,16 @@ import org.openjdk.jmh.annotations._
 import org.openjdk.jmh.annotations.Mode._
 
 import scala.scalanative.build._
+import scala.concurrent._
+import scala.concurrent.ExecutionContext.Implicits.global
+import scala.concurrent.duration._
 
 @Fork(1)
 @State(Scope.Benchmark)
 @BenchmarkMode(Array(AverageTime))
 @OutputTimeUnit(TimeUnit.MILLISECONDS)
-@Warmup(iterations = 5)
-@Measurement(iterations = 10)
+@Warmup(iterations = 5, time = 2, timeUnit = TimeUnit.SECONDS)
+@Measurement(iterations = 10, time = 2, timeUnit = TimeUnit.SECONDS)
 abstract class CodeGenBench(nativeConfig: NativeConfig => NativeConfig) {
   var config: Config = _
   var linked: linker.Result = _
@@ -32,7 +35,10 @@ abstract class CodeGenBench(nativeConfig: NativeConfig => NativeConfig) {
 
     val entries = build.ScalaNative.entries(config)
     util.Scope { implicit scope =>
-      linked = ScalaNative.link(config, entries)
+      linked = Await.result(
+        ScalaNative.link(config, entries),
+        Duration.Inf
+      )
     }
   }
 
@@ -49,7 +55,8 @@ abstract class CodeGenBench(nativeConfig: NativeConfig => NativeConfig) {
 
   @Benchmark
   def codeGen(): Unit = {
-    val paths = ScalaNative.codegen(config, linked)
+    val codegen = ScalaNative.codegen(config, linked)
+    val paths = Await.result(codegen, Duration.Inf)
     assert(paths.nonEmpty)
   }
 }