… tools-benchmarks/src/main/scala/scala/scalanative/benchmarks/testinterface/OptimizerBench.scala
… tools/src/main/scala/scala/scalanative/codegen/llvm/CodeGen.scala
… tools/src/main/scala/scala/scalanative/compat/CompatParColls.scala
… tools/src/test/scala/scala/scalanative/IncCompilationTest.scala
… tools/src/test/scala/scala/scalanative/LinkerSpec.scala
… tools/src/test/scala/scala/scalanative/OptimizerSpec.scala
… tools/src/test/scala/scala/scalanative/codegen/CodeGenSpec.scala
… tools/src/test/scala/scala/scalanative/linker/IssuesSpec.scala
… tools/src/test/scala/scala/scalanative/linker/ReachabilitySuite.scala