This change removes dependency on Scala parallel collections. These were removed from stdlib and were needed to be included as external dependency since 2.13. Fact of usage external dependencies is a blocker for #3233. The only remaining dependecy after this change would be ScalaTest.

Scala parallel collections have been replaced with direct usage of `Future`, especially `Future.traverse`.  However, this change requires us, to change the API of build. Now, `scalanative.Build.build` does require implicit `ExecutionContext` and does return `Future`. This now makes the API more similar to the API of Scala.js.  Because of viral nature of `scala.Future`, the `ExecutionContext` is propagated in all usages of code, that might execute in parallel.